requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
speedtest-cli==2.1.2
